Re-organize HIT related test cases
* Move HIT related test cases from test_scaling to test_slurm
* Rename HIT related test cases, test_hit_scaling -> test_slurm_scaling, test_hit_rare_failure -> test_error_handling
