+++ tests/integration-tests/configs/common/common.yaml
@@ -308,18 +308,6 @@ networking:
         oss: ["alinux2"]
         schedulers: ["slurm"]
 scaling:
-  test_scaling.py::test_hit_scaling:
-    dimensions:
-      - regions: ["us-west-1"]
-        instances: {{ common.INSTANCES_DEFAULT_X86 }}
-        oss: {{ common.OSS_ONE_PER_DISTRO }}
-        schedulers: ["slurm"]
-  test_scaling.py::test_hit_rare_failure:
-    dimensions:
-      - regions: ["ca-central-1"]
-        instances: {{ common.INSTANCES_DEFAULT_X86 }}
-        oss: ["alinux2"]
-        schedulers: ["slurm"]
   test_scaling.py::test_nodewatcher_terminates_failing_node:
     dimensions:
       - regions: ["sa-east-1"]
@@ -396,6 +384,18 @@ schedulers:
       - regions: ["ap-northeast-1"]
         instances: {{ common.INSTANCES_DEFAULT_ARM }}
         oss: {{ common.OSS_COMMERCIAL_ARM }}
+        schedulers: ["slurm"]
+  test_slurm.py::test_slurm_scaling:
+    dimensions:
+      - regions: ["us-west-1"]
+        instances: {{ common.INSTANCES_DEFAULT_X86 }}
+        oss: {{ common.OSS_ONE_PER_DISTRO }}
+        schedulers: ["slurm"]
+  test_slurm.py::test_error_handling:
+    dimensions:
+      - regions: ["ca-central-1"]
+        instances: {{ common.INSTANCES_DEFAULT_X86 }}
+        oss: ["alinux2"]
         schedulers: ["slurm"]
 spot:
   test_spot.py::test_spot_default:
