+++ tests/integration-tests/tests/schedulers/test_slurm.py
@@ -11,16 +11,34 @@
 # See the License for the specific language governing permissions and limitations under the License.
 import logging
 import time
+from datetime import datetime, timezone
 
+import boto3
 import pytest
 from assertpy import assert_that
 from remote_command_executor import RemoteCommandExecutionError, RemoteCommandExecutor
 from retrying import retry
 from time_utils import minutes, seconds
+from utils import get_compute_nodes_instance_ids
 
-from tests.common.assertions import assert_no_errors_in_logs, assert_no_node_in_ec2, assert_scaling_worked
+from tests.common.assertions import (
+    assert_errors_in_logs,
+    assert_no_errors_in_logs,
+    assert_no_node_in_ec2,
+    assert_num_instances_constant,
+    assert_num_instances_in_cluster,
+    assert_scaling_worked,
+    wait_for_num_instances_in_cluster,
+)
+from tests.common.hit_common import (
+    assert_compute_node_states,
+    assert_initial_conditions,
+    assert_num_nodes_in_scheduler,
+    submit_initial_job,
+    wait_for_num_nodes_in_scheduler,
+)
 from tests.common.mpi_common import compile_mpi_ring
-from tests.common.schedulers_common import SlurmCommands, TorqueCommands
+from tests.common.schedulers_common import SlurmCommands, TorqueCommands, get_scheduler_commands
 
 
 @pytest.mark.regions(["us-east-2"])
@@ -100,6 +118,429 @@ def test_slurm_pmix(pcluster_config_reader, clusters_factory):
     compile_mpi_ring(mpi_module, remote_command_executor, binary_path=binary_path)
     interactive_command = f"module load {mpi_module} && srun --mpi=pmix -N {num_computes} {binary_path}"
     remote_command_executor.run_remote_command(interactive_command)
+
+
+@pytest.mark.usefixtures("region", "os", "instance", "scheduler")
+@pytest.mark.slurm_scaling
+def test_slurm_scaling(scheduler, region, instance, pcluster_config_reader, clusters_factory, test_datadir):
+    """Test that slurm-specific scaling logic is behaving as expected for normal actions and failures."""
+    cluster_config = pcluster_config_reader(scaledown_idletime=3)
+    cluster = clusters_factory(cluster_config)
+    remote_command_executor = RemoteCommandExecutor(cluster)
+    scheduler_commands = get_scheduler_commands(scheduler, remote_command_executor)
+
+    _assert_cluster_initial_conditions(scheduler_commands, instance, 20, 20, 4, 1)
+    _test_partition_states(
+        scheduler_commands,
+        cluster.cfn_name,
+        region,
+        active_partition="ondemand1",
+        inactive_partition="ondemand2",
+        num_static_nodes=2,
+        num_dynamic_nodes=3,
+        dynamic_instance_type=instance,
+    )
+    _test_reset_terminated_nodes(
+        scheduler_commands,
+        cluster.cfn_name,
+        region,
+        partition="ondemand1",
+        num_static_nodes=2,
+        num_dynamic_nodes=3,
+        dynamic_instance_type=instance,
+    )
+    _test_replace_down_nodes(
+        remote_command_executor,
+        scheduler_commands,
+        test_datadir,
+        cluster.cfn_name,
+        region,
+        partition="ondemand1",
+        num_static_nodes=2,
+        num_dynamic_nodes=3,
+        dynamic_instance_type=instance,
+    )
+    _test_keep_or_replace_suspended_nodes(
+        scheduler_commands,
+        cluster.cfn_name,
+        region,
+        partition="ondemand1",
+        num_static_nodes=2,
+        num_dynamic_nodes=3,
+        dynamic_instance_type=instance,
+    )
+    assert_no_errors_in_logs(remote_command_executor, scheduler)
+
+
+@pytest.mark.usefixtures("region", "os", "instance", "scheduler")
+@pytest.mark.slurm_error_handling
+def test_error_handling(scheduler, region, instance, pcluster_config_reader, clusters_factory, test_datadir):
+    """Test that slurm-specific scaling logic can handle rare failures."""
+    cluster_config = pcluster_config_reader(scaledown_idletime=3)
+    cluster = clusters_factory(cluster_config)
+    remote_command_executor = RemoteCommandExecutor(cluster)
+    scheduler_commands = get_scheduler_commands(scheduler, remote_command_executor)
+
+    _assert_cluster_initial_conditions(scheduler_commands, instance, 10, 10, 1, 1)
+    _test_status_check_replacement(
+        remote_command_executor,
+        scheduler_commands,
+        cluster.cfn_name,
+        region,
+        partition="ondemand1",
+        num_static_nodes=1,
+        num_dynamic_nodes=1,
+        dynamic_instance_type=instance,
+    )
+    # Next test will introduce error in logs, assert no error now
+    assert_no_errors_in_logs(remote_command_executor, scheduler)
+    _test_clustermgtd_down_logic(
+        remote_command_executor,
+        scheduler_commands,
+        cluster.cfn_name,
+        region,
+        test_datadir,
+        partition="ondemand1",
+        num_static_nodes=1,
+        num_dynamic_nodes=1,
+        dynamic_instance_type=instance,
+    )
+
+
+def _assert_cluster_initial_conditions(
+    scheduler_commands,
+    instance,
+    expected_num_dummy,
+    expected_num_instance_node,
+    expected_num_static,
+    expected_num_dynamic,
+):
+    """Assert that expected nodes are in cluster."""
+    cluster_node_states = scheduler_commands.get_nodes_status()
+    c5l_nodes, instance_nodes, static_nodes, dynamic_nodes = [], [], [], []
+    logging.info(cluster_node_states)
+    for nodename, node_states in cluster_node_states.items():
+        if "c5l" in nodename:
+            c5l_nodes.append(nodename)
+        # "c5.xlarge"[: "c5.xlarge".index(".")+2].replace(".", "") = c5x
+        if instance[: instance.index(".") + 2].replace(".", "") in nodename:
+            instance_nodes.append(nodename)
+        if node_states == "idle":
+            if "-st-" in nodename:
+                static_nodes.append(nodename)
+            if "-dy-" in nodename:
+                dynamic_nodes.append(nodename)
+    assert_that(len(c5l_nodes)).is_equal_to(expected_num_dummy)
+    assert_that(len(instance_nodes)).is_equal_to(expected_num_instance_node)
+    assert_that(len(static_nodes)).is_equal_to(expected_num_static)
+    assert_that(len(dynamic_nodes)).is_equal_to(expected_num_dynamic)
+
+
+def _test_partition_states(
+    scheduler_commands,
+    cluster_name,
+    region,
+    active_partition,
+    inactive_partition,
+    num_static_nodes,
+    num_dynamic_nodes,
+    dynamic_instance_type,
+):
+    """Partition states INACTIVE and UP are processed."""
+    logging.info("Testing that INACTIVE partiton are cleaned up")
+    # submit job to inactive partition to scale up some dynamic nodes
+    init_job_id = submit_initial_job(
+        scheduler_commands,
+        "sleep 300",
+        inactive_partition,
+        dynamic_instance_type,
+        num_dynamic_nodes,
+        other_options="--no-requeue",
+    )
+    assert_initial_conditions(
+        scheduler_commands, num_static_nodes, num_dynamic_nodes, partition=inactive_partition, cancel_job_id=init_job_id
+    )
+    # set partition to inactive and wait for instances/node to terminate
+    scheduler_commands.set_partition_state(inactive_partition, "inactive")
+    # wait for all instances from inactive_partition to terminate
+    # active_partition should only have 2 static instances
+    wait_for_num_instances_in_cluster(cluster_name, region, 2)
+    # Assert no nodes in inactive partition
+    wait_for_num_nodes_in_scheduler(scheduler_commands, desired=0, filter_by_partition=inactive_partition)
+    # Assert active partition is not affected
+    assert_num_nodes_in_scheduler(scheduler_commands, desired=num_static_nodes, filter_by_partition=active_partition)
+    # set inactive partition back to active and wait for nodes to spin up
+    scheduler_commands.set_partition_state(inactive_partition, "up")
+    wait_for_num_nodes_in_scheduler(
+        scheduler_commands, desired=num_static_nodes, filter_by_partition=inactive_partition
+    )
+    # set inactive partition to inactive to save resources, this partition will not be used for later tests
+    scheduler_commands.set_partition_state(inactive_partition, "inactive")
+
+
+def _test_reset_terminated_nodes(
+    scheduler_commands, cluster_name, region, partition, num_static_nodes, num_dynamic_nodes, dynamic_instance_type
+):
+    """
+    Test that slurm nodes are reset if instances are terminated manually.
+
+    Static capacity should be replaced and dynamic capacity power saved.
+    """
+    logging.info("Testing that nodes are reset when instances are terminated manually")
+    init_job_id = submit_initial_job(
+        scheduler_commands,
+        "sleep 300",
+        partition,
+        dynamic_instance_type,
+        num_dynamic_nodes,
+        other_options="--no-requeue",
+    )
+    static_nodes, dynamic_nodes = assert_initial_conditions(
+        scheduler_commands, num_static_nodes, num_dynamic_nodes, partition, cancel_job_id=init_job_id
+    )
+    instance_ids = get_compute_nodes_instance_ids(cluster_name, region)
+    # terminate all instances manually
+    _terminate_nodes_manually(instance_ids, region)
+    # Assert that cluster replaced static node and reset dynamic nodes
+    _wait_for_node_reset(scheduler_commands, static_nodes, dynamic_nodes)
+    assert_num_instances_in_cluster(cluster_name, region, len(static_nodes))
+
+
+def _test_replace_down_nodes(
+    remote_command_executor,
+    scheduler_commands,
+    test_datadir,
+    cluster_name,
+    region,
+    partition,
+    num_static_nodes,
+    num_dynamic_nodes,
+    dynamic_instance_type,
+):
+    """Test that slurm nodes are replaced if nodes are marked DOWN."""
+    logging.info("Testing that nodes replaced when set to down state")
+    init_job_id = submit_initial_job(
+        scheduler_commands,
+        "sleep 300",
+        partition,
+        dynamic_instance_type,
+        num_dynamic_nodes,
+        other_options="--no-requeue",
+    )
+    static_nodes, dynamic_nodes = assert_initial_conditions(
+        scheduler_commands, num_static_nodes, num_dynamic_nodes, partition, cancel_job_id=init_job_id
+    )
+    # kill slurmd on static nodes, these nodes will be in down*
+    for node in static_nodes:
+        remote_command_executor.run_remote_script(str(test_datadir / "slurm_kill_slurmd_job.sh"), args=[node])
+    # set dynamic to down manually
+    _set_nodes_to_down_manually(scheduler_commands, dynamic_nodes)
+    _wait_for_node_reset(scheduler_commands, static_nodes, dynamic_nodes)
+    assert_num_instances_in_cluster(cluster_name, region, len(static_nodes))
+
+
+def _test_keep_or_replace_suspended_nodes(
+    scheduler_commands, cluster_name, region, partition, num_static_nodes, num_dynamic_nodes, dynamic_instance_type
+):
+    """Test keep DRAIN nodes if there is job running, or terminate if no job is running."""
+    logging.info(
+        "Testing that nodes are NOT terminated when set to suspend state and there is job running on the nodes"
+    )
+    job_id = submit_initial_job(
+        scheduler_commands,
+        "sleep 500",
+        partition,
+        dynamic_instance_type,
+        num_dynamic_nodes,
+        other_options="--no-requeue",
+    )
+    static_nodes, dynamic_nodes = assert_initial_conditions(
+        scheduler_commands, num_static_nodes, num_dynamic_nodes, partition
+    )
+    # Set all nodes to drain, static should be in DRAINED and dynamic in DRAINING
+    _set_nodes_to_suspend_state_manually(scheduler_commands, static_nodes + dynamic_nodes)
+    # Static nodes in DRAINED are immediately replaced
+    _wait_for_node_reset(scheduler_commands, static_nodes=static_nodes, dynamic_nodes=[])
+    # Assert dynamic nodes in DRAINING are not terminated during job run
+    _assert_nodes_not_terminated(scheduler_commands, dynamic_nodes)
+    # wait until the job is completed and check that the DRAINING dynamic nodes are then terminated
+    scheduler_commands.wait_job_completed(job_id)
+    scheduler_commands.assert_job_succeeded(job_id)
+    _wait_for_node_reset(scheduler_commands, static_nodes=[], dynamic_nodes=dynamic_nodes)
+    assert_num_instances_in_cluster(cluster_name, region, len(static_nodes))
+
+
+def _test_status_check_replacement(
+    remote_command_executor,
+    scheduler_commands,
+    cluster_name,
+    region,
+    partition,
+    num_static_nodes,
+    num_dynamic_nodes,
+    dynamic_instance_type,
+):
+    """Test nodes failing static check are correctly replaced."""
+    logging.info("Testing that nodes failing static check are correctly replaced")
+    job_id = submit_initial_job(
+        scheduler_commands,
+        "sleep 500",
+        partition,
+        dynamic_instance_type,
+        num_dynamic_nodes,
+        other_options="--no-requeue",
+    )
+    static_nodes, dynamic_nodes = assert_initial_conditions(
+        scheduler_commands, num_static_nodes, num_dynamic_nodes, partition, job_id
+    )
+    # Get network interface name from Head node, assuming Head node and Compute are of the same instance type
+    interface_name = remote_command_executor.run_remote_command(
+        "nmcli device status | grep ether | awk '{print $1}'"
+    ).stdout
+    logging.info("Detaching network interface {} on Compute nodes".format(interface_name))
+    # Only use dynamic nodes to test status check behavior
+    # Because static nodes will likely be placed in DOWN before status check failure shows in EC2
+    # Submit job that will detach network interface on all dynamic nodes, this will cause EC2 status check to fail
+    kill_job_id = scheduler_commands.submit_command_and_assert_job_accepted(
+        submit_command_args={
+            "command": "sudo ifconfig {} down && sleep 600".format(interface_name),
+            "partition": partition,
+            "constraint": "dynamic",
+            "other_options": "-a 1-{} --exclusive --no-requeue".format(num_dynamic_nodes),
+        }
+    )
+    # Assert that dynamic nodes with failing status check are put into DRAIN
+    # Can take up to 15 mins for status check to show
+    retry(wait_fixed=seconds(20), stop_max_delay=minutes(15))(assert_errors_in_logs)(
+        remote_command_executor,
+        ["/var/log/parallelcluster/clustermgtd"],
+        ["Setting nodes failing health check type ec2_health_check to DRAIN"],
+    )
+    scheduler_commands.cancel_job(kill_job_id)
+    # Assert dynamic nodes are reset
+    _wait_for_node_reset(scheduler_commands, static_nodes=[], dynamic_nodes=dynamic_nodes)
+    assert_num_instances_in_cluster(cluster_name, region, len(static_nodes))
+
+
+def _test_clustermgtd_down_logic(
+    remote_command_executor,
+    scheduler_commands,
+    cluster_name,
+    region,
+    test_datadir,
+    partition,
+    num_static_nodes,
+    num_dynamic_nodes,
+    dynamic_instance_type,
+):
+    """Test that computemgtd is able to shut nodes down when clustermgtd and slurmctld are offline."""
+    logging.info("Testing cluster protection logic when clustermgtd is down.")
+    submit_initial_job(
+        scheduler_commands,
+        "sleep infinity",
+        partition,
+        dynamic_instance_type,
+        num_dynamic_nodes,
+        other_options="--no-requeue",
+    )
+    static_nodes, dynamic_nodes = assert_initial_conditions(
+        scheduler_commands, num_static_nodes, num_dynamic_nodes, partition
+    )
+    logging.info("Killing clustermgtd and rewriting timestamp file to trigger timeout.")
+    remote_command_executor.run_remote_script(str(test_datadir / "slurm_kill_clustermgtd.sh"), run_as_root=True)
+    # Overwrite clusterctld heartbeat to trigger timeout path
+    timestamp_format = "%Y-%m-%d %H:%M:%S.%f%z"
+    overwrite_time_str = datetime(2020, 1, 1, tzinfo=timezone.utc).strftime(timestamp_format)
+    remote_command_executor.run_remote_command(
+        "echo -n '{}' | sudo tee /opt/slurm/etc/pcluster/.slurm_plugin/clustermgtd_heartbeat".format(overwrite_time_str)
+    )
+    # Test that computemgtd will terminate compute nodes that are down or in power_save
+    # Put first static node and first dynamic node into DOWN
+    # Put rest of dynamic nodes into POWER_DOWN
+    logging.info("Asserting that computemgtd will terminate nodes in DOWN or POWER_SAVE")
+    _set_nodes_to_down_manually(scheduler_commands, static_nodes[:1] + dynamic_nodes[:1])
+    _set_nodes_to_power_down_manually(scheduler_commands, dynamic_nodes[1:])
+    wait_for_num_instances_in_cluster(cluster_name, region, num_static_nodes - 1)
+
+    logging.info("Testing that ResumeProgram launches no instance when clustermgtd is down")
+    submit_initial_job(
+        scheduler_commands,
+        "sleep infinity",
+        partition,
+        dynamic_instance_type,
+        num_dynamic_nodes,
+    )
+
+    logging.info("Asserting that computemgtd is not self-terminating when slurmctld is up")
+    assert_num_instances_constant(cluster_name, region, desired=num_static_nodes - 1, timeout=2)
+
+    logging.info("Killing slurmctld")
+    remote_command_executor.run_remote_script(str(test_datadir / "slurm_kill_slurmctld.sh"), run_as_root=True)
+    logging.info("Waiting for computemgtd to self-terminate all instances")
+    wait_for_num_instances_in_cluster(cluster_name, region, 0)
+
+    assert_errors_in_logs(
+        remote_command_executor,
+        ["/var/log/parallelcluster/slurm_resume.log"],
+        ["No valid clustermgtd heartbeat detected"],
+    )
+
+
+def _wait_for_node_reset(scheduler_commands, static_nodes, dynamic_nodes):
+    """Wait for static and dynamic nodes to be reset."""
+    if static_nodes:
+        logging.info("Assert static nodes are placed in DOWN during replacement")
+        # DRAIN+DOWN = drained
+        _wait_for_compute_nodes_states(
+            scheduler_commands, static_nodes, expected_states=["down", "down*", "drained", "drained*"]
+        )
+        logging.info("Assert static nodes are replaced")
+        _wait_for_compute_nodes_states(scheduler_commands, static_nodes, expected_states=["idle"])
+    # dynamic nodes are power saved after SuspendTimeout. static_nodes must be checked first
+    if dynamic_nodes:
+        logging.info("Assert dynamic nodes are power saved")
+        _wait_for_compute_nodes_states(scheduler_commands, dynamic_nodes, expected_states=["idle~"])
+
+
+def _assert_nodes_not_terminated(scheduler_commands, nodes, timeout=5):
+    logging.info("Waiting for cluster daemon action")
+    start_time = time.time()
+    while time.time() < start_time + 60 * (timeout):
+        assert_that(set(nodes) <= set(scheduler_commands.get_compute_nodes())).is_true()
+        time.sleep(20)
+
+
+def _set_nodes_to_suspend_state_manually(scheduler_commands, compute_nodes):
+    scheduler_commands.set_nodes_state(compute_nodes, state="drain")
+    # draining means that there is job currently running on the node
+    # drained would mean we placed node in drain when there is no job running on the node
+    assert_compute_node_states(scheduler_commands, compute_nodes, expected_states=["draining", "drained"])
+
+
+def _set_nodes_to_down_manually(scheduler_commands, compute_nodes):
+    scheduler_commands.set_nodes_state(compute_nodes, state="down")
+    assert_compute_node_states(scheduler_commands, compute_nodes, expected_states=["down"])
+
+
+def _set_nodes_to_power_down_manually(scheduler_commands, compute_nodes):
+    scheduler_commands.set_nodes_state(compute_nodes, state="power_down")
+    time.sleep(5)
+    scheduler_commands.set_nodes_state(compute_nodes, state="resume")
+    assert_compute_node_states(scheduler_commands, compute_nodes, expected_states=["idle~"])
+
+
+@retry(wait_fixed=seconds(20), stop_max_delay=minutes(5))
+def _wait_for_compute_nodes_states(scheduler_commands, compute_nodes, expected_states):
+    assert_compute_node_states(scheduler_commands, compute_nodes, expected_states)
+
+
+def _terminate_nodes_manually(instance_ids, region):
+    ec2_client = boto3.client("ec2", region_name=region)
+    for instance_id in instance_ids:
+        instance_states = ec2_client.terminate_instances(InstanceIds=[instance_id]).get("TerminatingInstances")[0]
+        assert_that(instance_states.get("InstanceId")).is_equal_to(instance_id)
+        assert_that(instance_states.get("CurrentState").get("Name")).is_in("shutting-down", "terminated")
+    logging.info("Terminated nodes: {}".format(instance_ids))
 
 
 def _test_mpi_job_termination(remote_command_executor, test_datadir):
