+++ tests/integration-tests/tests/schedulers/test_slurm/test_error_handling/slurm_kill_slurmctld.sh
