+++ tests/integration-tests/tests/schedulers/test_slurm/test_slurm_scaling/slurm_kill_slurmd_job.sh
